pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer is a website that analyzes specified pages for SEO suitability"
authors = ["BogdanBarylo <barylob@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0"
gunicorn = "20.1.0"
python-dotenv = "^1.0"
psycopg2-binary = "^2.9"
jinja2 = "^3.1"
requests = "^2.31"
validators = "^0.22"
beautifulsoup4 = "^4.12"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
